mesa: fix darwin and use meson auto features
This fixes the darwin build, while also using Meson’s auto features as
much as possible. As a result, we avoid using having to specify
default drivers and instead delegate that to Mesa’s build system.
Removed other flags that were specified to the default in Mesa.

The -fno-common is needed to address undefined symbol _lp_dummy_tile
in the build.