systemd: point to commit id instead of branch name
branch names are mutable, and with
https://github.com/NixOS/systemd/pull/29 being merged in, the nixos-v242
branch advanced from 5c20aab77900f478fd380ab189787d80e4a35963 to
40eb070cb309ec09def0ecdeaf7514c702200835, causing systemd's
fetchFromGitHub to fail with a sha256sum mismatch (when not relying on
the cache).

Fix this, by pointing systemd.src to the commit id before the branch
advancement. This won't cause a rebuild, as the sha256 stayed the same.

Fast-forwarding systemd to 40eb070cb309ec09def0ecdeaf7514c702200835 will
be done in https://github.com/NixOS/nixpkgs/pull/63784 , which also uses
the commit id, and not a branch name for rev.