hplip: retain udev rules in WithPlugin version
Beginning with commit 8ffd6af27e6dfae0a3b4d77db03409f62c5b8d98,
the build recipe for `hplipWithPlugin` explicitely drops a
udev rule file that is delivered with the sources.
In `hplip` (without plugin), the rule file is retained.

The udev rules takes care of uploading firmware
to certain printers (and maybe of other things),
e.g. to HP LaserJet 1000.

This commit reinstates the rule file and
expands the fixup script so that the rules
are also patched in the WithPlugin version.