lib/modules: add mkUnset, mkOverrideUnset
Using only `mkForce`/`mkOverride`, a module or configuration has no way
of removing an attribute specified elsewhere; the best you could do is
to replace that attribute with a null, empty set, or similar. This may
result in undesired behavior, if a module implementation takes some
action for every attribute in a set, for example.

`mkOverrideUnset` is a function intended to be used like `mkOverride`: a
`config.some.attribute.path = lib.mkOverrideUnset p` means that, if `p`
is the winning priority among all definitions of `some.attribute.path`,
then `some.attribute.path` will not exist at all in the final
configuration values.

`mkUnset` is a shorthand for `mkOverrideUnset 50`, analogous to
`mkForce`.