way-cooler: Mark as broken
The project is in full-rewrite state, which also breaks backwards
compatibility.
Right now, the project README warns that way-cooler is not usable in its
current state.

Thus mark this as broken, so die-hard users can still use it.

This patch should be reverted as soon as there is a new release.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>