ghc: Don't strip compilers
To test this I built a version of the compiler with `dontStrip = True`
and the vanilla version. The size of the result is both 1.4gb which
indicates that the stripping doesn't do anything meaningful.

Not stripping means that the debug rts is properly packages as it
contains DWARF information and unused debugging symbols.

Fixes #63511