WIP: build-rust-crate: fix for binary crate building.
See #64313.

build-rust-crate used to always build a crate that had a source file
with the same name as the crate as a library crate. This would break in
the cases of the crate being a binary crate containing a source file
with the same name, e.g. in xargo.

Cargo used to erroneously allow a crate with the following contents in it's cargo
toml:
```toml
[lib]
name = "foo"
```
having its main source file named `src/foo.rs`. This is behaviour that
cargo warns about.

This pull request is marked as WIP because the version of way-cooler
that is in nixpkgs depends on an old version of the crate `c_vec` that
assumed the erroneous behaviour of cargo. Unfortunately this pull
request will make the build of way-cooler fail until the dependency is
updated to a new version.