nixos/systemd-confinement: Fix ProtectStrict=full support
Problem was that if you enabled DynamicUser, it would remount
/ as read-only which wouldn't work, as we were using
the fact that / was read-write to set up all the mounts.

We work around this by actually making sure all the directories
that will be mount points exist apriori.  For things
for which mount points are dynamically generated on the fly,
we use TemporaryFileSystem (cache, log, state).

We also use TemporaryFileSystm for /nix/store. However,
we might not actually need it, as we know all the mount points
apriori.

TODO:

* See how we interact with APIVFS
* See if we can get rid of the TemporaryFileSystem for /nix/store,
because currently /nix/store _itself_ is read/write in the container (it
already was) and that's surely not something we want!