nixos/activation: no output for empty snippets
There are several empty snippets in a standard configuration. They are
present only for backward compatibility with other snippets which may
have declared dependencies on them.

This patch eliminates the per-snippet boilerplate for empty snippets.
If, someday, we get to the point where there are no non-empty activation
snippets, then this will lead to an empty string where the snippets used
to be.

Note that I didn't want newlines inserted where empty scripts used to be
either, so rather than using textClosureMap with its built-in "\n"
separators, I switched to using the underlying textClosureList, and
adding an extra newline only for non-empty snippets.