nixos/users-groups: create homedirs in users snippet
In a subsequent commit I'll eliminate update-users-groups in the case
where the contents of /etc/passwd and /etc/group can be precomputed at
build time. But that requires that either there are no createHome users
or that creating homedirs is done by something other than that script.
So in this commit I've implemented the latter approach.

Tested by instantiating a minimal configuration.nix with:

    boot.isContainer = true;
    services.mingetty.autologinUser = "root";

both with and without a `isNormalUser = true` user.

Without any createHome users, the "users" snippet is empty so it does
not show up in the activation script and /home does not get created.

With a normal user (which implies createHome), the "users" snippet
appears in the activation script, and in a container which did not
previously contain /home, the snippet successfully creates /home and the
user's homedir under it. After changing the mode and ownership of the
user's homedir, re-running the activation script preserves the new mode
but resets the ownership, which is consistent with the perl version.