nixos/etc: defer setting mode/ownership to end
If environment.etc is being used to set up /etc/passwd or /etc/group,
then we can't rely on getpwnam and getgrnam until we've already put all
the new /etc files in place.

The easiest way to fix that is to make all the symlinks and do all the
copies first, and then go back to each copied file and fix its ownership
and permissions.

Since we're deferring ownership changes, we have to also defer
permissions changes; otherwise an unprivileged service could lose access
to its config files during this window. Leaving files readable by
everyone for a bit isn't a security problem because the source of these
files is already world-readable in the Nix store.