nixos/users-groups: split users snippet around etc
In order to create /etc/passwd and /etc/group via environment.etc, the
"etc" snippet becomes a dependency of "users", which is the snippet that
other snippets depend on. But when "users" creates those files instead,
"etc" needs to depend on "users".

This patch splits "users" into two snippets:

- users-before-etc does everything users currently does, and etc depends
  on it.
- users is now an empty snippet which exists only to support ordering
  constraints for other snippets. It depends on etc.

This resolves the potential circular dependency.