nixos/users-groups: fix handling of numeric groups
Although this feature isn't documented, it looks like
update-users-groups.pl is supposed to recognize a user's primary group
"name" that's composed entirely of digits as a GID instead.

However it did this only for single-digit GIDs, which I can't imagine is
what was intended.

So this patch generalizes it to accept GIDs of any length.