types: Fix path.check for cleanSourceWith output
Advantage of `cleanSourceWith` is that it composes without copying intermediate output to the store.  Currently using `path.check` breaks this.  Consider:

```nix
let cleanedSource1 = cleanSourceWith { src = ./.; filter = filter1; };
in
  assert (path.check cleanedSource1);
  cleanSourceWith { src = cleanedSource1; filter = filter2; }
```

Currently this will copy all the files that pass filter1 to the store in order to check that the resulting path starts with "/".

Luckily `cleanSourceWith` exposes a `origSrc` property that we can check.