fish: use internal wcwidth
Fish 3.0 has an updated an more robust handling of unicode glyphs. Per
the original author of the INTERNAL_WCWIDTH flag, it was something of
misfeature, and they regret that NixOS came to rely on it.

Removes the flag from the Nix expression.

Flag was added originally to Nixpkgs in 68076b7d4938ce788dc9eaaf8789dfccc30e2ca2

It is being removed entirely from upstream fish:
https://github.com/fish-shell/fish-shell/pull/5777.