wire-desktop: Add darwin support
Wire Desktop is available for linux, mac os and windows. I figured
adding darwin support would be cromulent. Note that the versions don't
align 100%, this is how it's released upstream.

I refactored the derivation to seperate all the linux-specific parts. I
also sorted the dependencies and grouped them.

The changes were based on the derivation for electron. I changed the
construction from calling `mkDerivation` on a conditional merger of two
sets by moving the `mkDerivation` calls into the conditional and up to
the local bindings for `linux` and `darwin`. This required moving
`pname` and `meta` up to local bindings.