nixos: only trigger restarts on /etc/hosts if present
On a system with nss-myhostname configured, as it is by default on
NixOS, /etc/hosts is not necessary for most people. Although NixOS
currently creates it unconditionally, this patch allows for setting
`environment.etc.hosts.enable = false`, and also clears the way for
NixOS to someday not create the file in the first place.