boot tests: fix globbing in qemu configuration
Before the refactoring for IPXE the configuration was literal perl
code and included a call to `glob`. After refactoring it became a map
of strings, where those strings contained the old perl
code. Ultimately this was passing perl fragements to the shell
interpreting the qemu command line. To restore the globbing behaviour,
we can inject the equivalent shell glob.