Resolve symlinks before trying to recognize a lisp implementation
The CCL package installs a symlink named "ccl" that points at the
actual implementation executable: lx86cl64 (or lx86cl for 32 bit).
When clwrapper is used with CCL as the backing implementation, this
script fails to recognize the implementation.  By resolving the
symlink, we are able to recognize which implementation we're actually
working with.