Use if instead of && for deciding whether to run a command
The only difference between these forms is the return value when
"$NIX_LISP_SKIP_CODE" is the empty string.  In the original
formulation, the script would return a false exit status.  In the new
formulation, it will return a true exit status.

Its useful to be able to source cl-wrapper.sh (to get the variables it
establishes), and its a bit annoying that sourcing it with
NIX_LISP_SKIP_CODE=1 results in a false exit status.