bazel-deps: 2019-02-01 -> 2019-07-11, mark as broken
Also drop preInstall cleanup for dependencies. The reason is while it's more
thorough than default cleanup in buildBazelPackage if such a problem happens we
should fix buildBazelPackage instead. Perhaps even move this (awesome!) snippet
there but it's very slow-running so we'd rather attempt to fix it in other
ways.

Anyway after an update at least .deps build, checked with `nix-build -A --check`.