firejail: allow wrapping .desktop applications
In addition to wrapping binaries it's convenient to wrap the 'Exec' attribute in
.desktop files with firejail. For example, wrapping the firefox binary with firejail
removes the firefox.desktop file so you cannot start firefox using an
application launches like in Gnome.

This patch allows to specify an attribute set of applications whose desktop file
to wrap. For convenience it allows wrapping their binary through the same config
element. Since firejail options cannot be part of the desktop file attribute, a
separate option is provided for those.