nixos/users-groups: new users.enforceStaticIds asserts
Per review comments from @arcnmx and @Infinisil on #64594, don't use
auto-detection to determine whether to prebuild /etc/passwd and
/etc/group.

Instead offer a new config option, by analogy with
users.enforceIdUniqueness: users.enforceStaticIds provides an assertion
that all users and groups have an ID statically assigned.

If this option is turned on and the assertions pass, _and_
users.mutableUsers is turned off, then /etc/passwd and /etc/group will
be generated at configuration build time instead of activation time.

This allows admins to declare their intent and get a build failure if
that intent is violated.

Adding a configuration option also provided a natural place to document
this feature (thanks again to @arcnmx for pointing that out) so this
commit adds some documentation accordingly.

I didn't want to rebase this PR but I've tested this commit both as-is
and after a merge from current git master, by instantiating a minimal
configuration.nix with several variations on:

    boot.isContainer = true;
    users.mutableUsers = false;
    users.enforceStaticIds = true;
    users.users.root.initialHashedPassword = "";

On this commit, the new assertions fire unless this option is also set:

    services.nscd.enable = false;

Since #64268 has been merged to master, after merging master onto this
branch, that option is no longer necessary.