tdesktopPackages.*: Remove the attributes in favour of tdesktop
This is a refactoring of "tdesktop", i.e. the resulting derivation isn't
affected by this commit (only the tdesktopPackages.* attributes are
removed).

I decided to remove the attributes "tdesktopPackages.*" as I don't
maintain the preview version anymore. There are regular stable releases
and we depend on the patches from Arch Linux which only track the stable
version as well, i.e. I would have to maintain our own patches, which
would either require an additional repository or "bloat" Nixpkgs (apart
from the fact that it would be unnecessarily time-consuming).