nixos/openvpn: user, forwarding and local state dir
We change a few things here:

a) Create the local openvpn user and group so that openvpn can drop privileges.
This is not switched on by default.

b) Define an option to enable IP forwarding which would have to be defined
outside of the openvpn module.

c) Create a local state and runtime directories for anything openvpn might create (replay
persistence logs, current connections, etc).