kernel/modules-closure.sh: fix the modules closure
Prior to this commit the modules closure could reference nonsensical
pieces of data in to form of module parameters. These were treated
as seperate modules.

"modprobe --show-depends" outputs a line of the form
"insmod module_name module_args", we are only interested in the
module name, so we extract it with sed.

Also fixes a bug in the original implementation where the check for missing modules did not work as indended (the exit status of
modprobe was ignored due to a pipe).