ffmpeg_4: Enable support for AV1 decoding via dav1d by default
This is e.g. required for mpv (depends on ffmpeg_4) to play AV1 videos.
Fixes #54990.

But since dav1d is only a AV1 decoder this doesn't support AV1 encoding
as well (that would require an additional dependency on libaom).
The dependency on dav1d can be disabled by overriding it to null.