buildPerlPackage: phase out name
Some time ago `perlPackages.*` have been refactored to use `buildPerlPackage` with `pname` and `version` instead of `name`
and a warning was printed when it was used with `name` to prompt users fix their code.
As time passed, now the warning is being replaced with a fatal error and the support for `name` is removed.